# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TripJar API - collaborative group savings platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jars:
    get:
      operationId: listJars
      tags: [jars]
      summary: List all jars for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JarSummary"
    post:
      operationId: createJar
      tags: [jars]
      summary: Create a new jar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJarRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jar"
  /jars/{jarId}:
    get:
      operationId: getJar
      tags: [jars]
      summary: Get jar details
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jar"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJar
      tags: [jars]
      summary: Update jar
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJarRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jar"
  /jars/{jarId}/launch:
    post:
      operationId: launchJar
      tags: [jars]
      summary: Launch a jar (change status from draft to saving)
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jar"
  /jars/{jarId}/cancel:
    post:
      operationId: cancelJar
      tags: [jars]
      summary: Cancel a jar
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelJarRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jar"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Jar:
      type: object
      properties:
        id:
          type: string
        organizerId:
          type: string
        name:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [Vacation, Cruise, Wedding, Honeymoon, MissionTrip, Reunion, Celebration, Other]
        description:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        targetDate:
          type: string
          format: date
        goalAmountCents:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [Draft, Inviting, Saving, CommitmentPending, Committed, FullyFunded, Completed, Cancelled]
        approvalThreshold:
          type: number
        memberCount:
          type: integer
        totalSavedCents:
          type: integer
        percentFunded:
          type: number
        daysRemaining:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, organizerId, name, slug, category, targetDate, goalAmountCents, currency, status, approvalThreshold, memberCount, totalSavedCents, percentFunded, createdAt, updatedAt]
    JarSummary:
      type: object
      properties:
        id:
          type: string
        organizerId:
          type: string
        name:
          type: string
        category:
          type: string
        destination:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        targetDate:
          type: string
          format: date
        goalAmountCents:
          type: integer
        currency:
          type: string
        status:
          type: string
        memberCount:
          type: integer
        totalSavedCents:
          type: integer
        percentFunded:
          type: number
        daysRemaining:
          type: integer
          nullable: true
        userRole:
          type: string
          nullable: true
        health:
          $ref: "#/components/schemas/JarHealth"
          nullable: true
      required: [id, organizerId, name, category, targetDate, goalAmountCents, currency, status, memberCount, totalSavedCents, percentFunded]
    CreateJarRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [Vacation, Cruise, Wedding, Honeymoon, MissionTrip, Reunion, Celebration, Other]
        description:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        targetDate:
          type: string
          format: date
        goalAmountCents:
          type: integer
        currency:
          type: string
          default: USD
        approvalThreshold:
          type: number
          default: 0.67
      required: [name, category, targetDate, goalAmountCents]
    UpdateJarRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        destination:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        targetDate:
          type: string
          format: date
        goalAmountCents:
          type: integer
        approvalThreshold:
          type: number
    CancelJarRequest:
      type: object
      properties:
        reason:
          type: string
    JarHealth:
      type: object
      properties:
        status:
          type: string
          enum: [Ahead, OnTrack, NeedsAttention, AtRisk]
        message:
          type: string
        actualPercentage:
          type: number
        expectedPercentage:
          type: number
        daysElapsed:
          type: integer
        totalDays:
          type: integer
        amountBehindCents:
          type: integer
        estimatedCompletionDate:
          type: string
          format: date
          nullable: true
      required: [status, message, actualPercentage, expectedPercentage, daysElapsed, totalDays, amountBehindCents]

    # MEMBERS
...